Change lua print function to write in a file
This lets you retrieve print messages done by the test binding
in a file log in addition of binder runtime log messages.
By default, the standard output is redirected to a file
in the workdir of the binder.

This is convenient when you have a heavy loaded binder output. It's
more difficult to extract the test results.

Change-Id: Ifae540652ae431d4d9703a19673e5bc69c2b315f
Signed-off-by: Romain Forlot <romain.forlot@iot.bzh>