Fix: events callback aren't correctly registered
They were stored in the mapis sections instead of the events
one.

Clean and handle raw events as well. Before that it could
only process events coming from monitoring api. Now it could
receive the event directly.

Make sure that the event handle is valid before sends the
request response. This prevent to subcribe or push an event
while the event handle hasn't been correctly created. This
could happens if the testVerb hasn't not been called from a
session. In that case, the response will just be that tests
has been launched.

Format.

Change-Id: I0aa522939162684f91dd426cc14919bb0ec3f69e
Signed-off-by: Romain Forlot <romain.forlot@iot.bzh>