Make afTest test itself correctly
For native build and test purposes you need to start 2
bindings to be able to test correctly the afb-test binding.
The first one will answer on the main API name afTest and the
second to launch the test.
This is the closest test case similar to a test on a target
board.

Change-Id: I6b961c0e41cdd3eef387b7f5156ed334e8f6c3de
Signed-off-by: Romain Forlot <romain.forlot@iot.bzh>