Handle event in core binding
Prefers to handle the events reception by the binding, in C, rather than using
the LUA interpreter to avoid race condition on the LUA context. Because we are
waiting events in LUA context using binder call sync, when the event is received
then 2 threads operate simultaneously on LUA context, the waiting event thread and
the receiving event thread.

Bug-AGL: SPEC-2135

Change-Id: Ied0a78a61263b8fd41305969c636a491c6bb0295
Signed-off-by: Romain Forlot <romain.forlot@iot.bzh>