Fix signal 11 segfault in pthread
This fix a segfault happening using systemd service where the the end of test from the
binder point of view is done but the test are running. This was because the binding answer
to the binder's request at the beginning of the test instead of the end.

This commit adds also useful details in the request answer

Change-Id: I8dd5b0e7a61608d3b7736ed278d92376f0d11fd3
Signed-off-by: Romain Forlot <romain.forlot@iot.bzh>